feat:  updating the edge-weights-pop-up
