+++ index.html
@@ -71,10 +71,6 @@
         </div>
     </div>
 
-    <div class="edge-popup-container">
-
-    </div>
-
     <ul id="context-menu" class="context-menu">
         <li id="move" class="context-menu_item">
             Move Node
@@ -276,18 +272,20 @@
 
 <!--pop-ups -->
 <script src="js/Canvas/Pop-Up/Modal.js" type="module"></script>
+<script src="js/Canvas/Pop-Up/weighted-edge-input-pop-up.js" type="module"></script>
+<script src="js/Canvas/Pop-Up/edge-weights-pop-up.js" type="module"></script>
 <script src="js/sidebar/buttons/matrix-input.js" type="module"></script>
 <script src="js/sidebar/buttons/textInput.js" type="module"></script>
 
 <script src="js/Canvas/ContextMenu/conterxt-menu.js"></script>
 <script src="js/Canvas/edges-info.js"></script>
-<script src="js/Canvas/edge-clicked-handler.js"></script>
+<script src="js/Canvas/edge-clicked-handler.js" type="module"></script>
 <script src="js/Canvas/Graph.js"></script>
 <script src="js/UI.js"></script>
 <script src="js/functions.js"></script>
 <script src="js/Canvas/canvasFunctions.js"></script>
 <script src="js/index.js"></script>
-<script src="js/listeners.js"></script>
+<script src="js/listeners.js" type="module"></script>
 
 <script src="js/Algorithms/Draw-as-tree.js"></script>
 <script src="js/sidebar/toolbox.js" type="module"></script>
